+++ src/main/java/kitchenpos/table/application/validator/TableChangeEmptyValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.application;
+package kitchenpos.table.application.validator;
 
 public interface TableChangeEmptyValidator {
