+++ src/main/java/kitchenpos/table/application/validator_impl/TableGroupTableValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.application.validator;
+package kitchenpos.table.application.validator_impl;
 
 import static kitchenpos.table_group.domain.exception.TableGroupExceptionType.ORDER_TABLE_IS_NOT_EMPTY;
 import static kitchenpos.table_group.domain.exception.TableGroupExceptionType.ORDER_TABLE_IS_NOT_PRESENT_ALL;
